Moved inventory module to reservations
