+++ apps/api/src/reservation/application/inventory.service.integration.spec.ts
@@ -6,8 +6,8 @@ import { DatabaseModule } from 'shared/database/database.module';
 import { ConfigModule } from '@nestjs/config';
 import databaseConfig from 'config/database.config';
 import { InventoryService } from './inventory.service';
-import { InventoryItem } from './inventory-item.entity';
-import { Location } from './location.entity';
+import { InventoryItem } from '../core/entities/inventory-item.entity';
+import { Location } from '../core/entities/location.entity';
 import { createInventoryItem } from 'shared/database/factories/inventory-item.factory';
 import { createLocation } from 'shared/database/factories/location.factory';
 import { createReservation } from 'shared/database/factories/reservation.factory';
