+++ apps/api/src/reservation/core/entities/inventory-item.entity.ts
