+++ apps/api/src/reservation/core/entities/location.entity.ts
